Rename app package
From com.voidzen2fa to com.voidzen.mfa
