+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.voidzen2fa",
+    package = "com.voidzen.mfa",
 )
 
 android_resource(
     name = "res",
-    package = "com.voidzen2fa",
+    package = "com.voidzen.mfa",
     res = "src/main/res",
 )
