+++ android/app/src/main/java/com/voidzen/mfa/MainApplication.java
@@ -1,4 +1,4 @@
-package com.voidzen2fa;
+package com.voidzen.mfa;
 
 import android.app.Application;
 import android.content.Context;
@@ -9,7 +9,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.config.ReactFeatureFlags;
 import com.facebook.soloader.SoLoader;
-import com.voidzen2fa.newarchitecture.MainApplicationReactNativeHost;
+import com.mfa.newarchitecture.MainApplicationReactNativeHost;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 
@@ -73,7 +73,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.voidzen2fa.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.com.voidzen.mfa.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
